使用zuul搭建api-gateway、添加user服务、bug fix

+++ order/order-server/src/main/java/cn/fulgens/order/dto/OrderDto.java
@@ -42,12 +42,12 @@ public class OrderDto {
     /**
      * 订单状态, 默认为新下单
      */
-    private Byte orderStatus;
+    private Integer orderStatus;
 
     /**
      * 支付状态, 默认未支付
      */
-    private Byte payStatus;
+    private Integer payStatus;
 
     /**
      * 订单项集合
